Move post funcs from cpp_utils to cpp_hooks
